pyproject.toml
[project]
name = "whisperstt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.2.6",
    "openai-whisper",
    "scipy>=1.15.3",
    "sounddevice>=0.5.2",
    "soundfile>=0.13.1",
    "streamlit>=1.45.1",
    "streamlit-webrtc>=0.62.4",
    "torch>=2.7.0",
]

[tool.uv.sources]
openai-whisper = { git = "https://github.com/openai/whisper.git" }
